requirements.txt
asgiref==3.10.0
asttokens==3.0.0
cbor2==5.6.5
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
immutables==0.21
lark==1.2.2
packaging==25.0
pillow==12.0.0
pycryptodome==3.23.0
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vyper==0.4.3
wheel==0.45.1
